+++ lib/src/utils/places_operation.dart
@@ -1,0 +1,9 @@
+enum PlacesOperation {
+  autocomplete,
+  details,
+  findPlace,
+  nearbySearch,
+  textSearch,
+  photo,
+  queryAutocomplete,
+}